Rework the handling of players and the console calling the /sort command.
With these changes, most of the methods of the Player interface is delegated to the actual player. When run from the server's console, the CommandSender is implemented by the console and most player related functionality is stub-ed. A few methods are implemented as they are known to be called when handling items in an inventory by other plugins.